requirements.txt
pipdeptree==0.13.2
pkg-resources==0.0.0
pycryptodome==3.9.0
